(maint) Switch from int to long
When testing usage of nuget.org, there is a single package that has
over 3 billion downloads.  When attempting to view this package in
Chocolatey GUI, there is a convert error, when it tries to convert this
nubmer into an int.  This is a similar change to what was done in CCR,
when we crossed over the 1B mark for downloads.

+++ Source/ChocolateyGui.Common/Models/Package.cs
@@ -20,7 +20,7 @@ namespace ChocolateyGui.Common.Models
 
         public string Description { get; set; }
 
-        public int DownloadCount { get; set; }
+        public long DownloadCount { get; set; }
 
         public string GalleryDetailsUrl { get; set; }
 
@@ -68,6 +68,6 @@ namespace ChocolateyGui.Common.Models
 
         public NuGetVersion Version { get; set; }
 
-        public int VersionDownloadCount { get; set; }
+        public long VersionDownloadCount { get; set; }
     }
 }
